admin delete validation was missing, fixed made

+++ controllers/profiles/doctor.js
@@ -46,11 +46,15 @@ router.put("/:userId/doctors/:id", async (req, res) => {
 
 // DELETE DOCTOR
 router.delete("/:userId/doctors/:id", async (req, res) => {
-  try {
-    const doctor = await Doctor.findByIdAndDelete(req.params.id);
-    res.json({ message: "Doctor Deleted" }, doctor);
-  } catch (error) {
-    res.status(404).json({ error: error.message });
+  if (req.user.type[2000]) {
+    try {
+      const doctor = await Doctor.findByIdAndDelete(req.params.id);
+      res.json({ message: "Doctor Deleted" }, doctor);
+    } catch (error) {
+      res.status(404).json({ error: error.message });
+    }
+  } else {
+    res.status(404).json({ error: "Oops, something went wrong" });
   }
 });
